+++ chains/ethereum/listener_test.go
@@ -11,8 +11,8 @@ import (
 	"time"
 
 	msg "github.com/ChainSafe/ChainBridge/message"
-	utils "github.com/ChainSafe/ChainBridge/utils/ethereum"
-	ethtest "github.com/ChainSafe/ChainBridge/utils/ethereum/testing"
+	utils "github.com/ChainSafe/ChainBridge/shared/ethereum"
+	ethtest "github.com/ChainSafe/ChainBridge/shared/ethereum/testing"
 	"github.com/ethereum/go-ethereum/common"
 	ethcrypto "github.com/ethereum/go-ethereum/crypto"
 )
