+++ chains/ethereum/test_utils_test.go
@@ -8,9 +8,14 @@ import (
 	"testing"
 	"time"
 
+	"github.com/ChainSafe/ChainBridge/bindings/Bridge"
 	"github.com/ChainSafe/ChainBridge/keystore"
 	msg "github.com/ChainSafe/ChainBridge/message"
+	utils "github.com/ChainSafe/ChainBridge/shared/ethereum"
 	"github.com/ChainSafe/log15"
+	"github.com/ethereum/go-ethereum/accounts/abi/bind"
+	"github.com/ethereum/go-ethereum/common"
+	"github.com/ethereum/go-ethereum/common/hexutil"
 )
 
 const TestEndpoint = "ws://localhost:8545"
@@ -21,9 +26,7 @@ var TestTimeout = time.Second * 10
 var AliceKp = keystore.TestKeyRing.EthereumKeys[keystore.AliceKey]
 var BobKp = keystore.TestKeyRing.EthereumKeys[keystore.BobKey]
 
-var TestNumRelayers = 2
 var TestRelayerThreshold = big.NewInt(2)
-var TestMintAmount = big.NewInt(1000)
 
 var aliceTestConfig = &Config{
 	id:       msg.ChainId(0),
@@ -58,4 +61,42 @@ func newLocalConnection(t *testing.T, cfg *Config) *Connection {
 	}
 
 	return conn
+}
+
+func deployTestContracts(t *testing.T, id msg.ChainId) *utils.DeployedContracts {
+	contracts, err := utils.DeployContracts(
+		hexutil.Encode(AliceKp.Encode())[2:],
+		uint8(id),
+		TestEndpoint,
+		TestRelayerThreshold,
+	)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	return contracts
+}
+
+// createErc20Deposit deploys a new erc20 token contract mints, the sender (based on value), and creates a deposit
+func createErc20Deposit(contract *Bridge.Bridge,
+	txOpts *bind.TransactOpts,
+	rId msg.ResourceId,
+	originHandler,
+	destRecipient common.Address,
+	destId msg.ChainId,
+	amount *big.Int) error {
+
+	data := utils.ConstructErc20DepositData(rId, destRecipient.Bytes(), amount)
+
+	// Incrememnt Nonce by one
+	txOpts.Nonce = txOpts.Nonce.Add(txOpts.Nonce, big.NewInt(1))
+	if _, err := contract.Deposit(
+		txOpts,
+		uint8(destId),
+		originHandler,
+		data,
+	); err != nil {
+		return err
+	}
+	return nil
 }
