+++ chains/substrate/connection_test.go
@@ -6,7 +6,7 @@ package substrate
 import (
 	"testing"
 
-	utils "github.com/ChainSafe/ChainBridge/utils/substrate"
+	utils "github.com/ChainSafe/ChainBridge/shared/substrate"
 	"github.com/centrifuge/go-substrate-rpc-client/types"
 )
