+++ chains/substrate/writer_test.go
@@ -9,8 +9,8 @@ import (
 	"testing"
 
 	message "github.com/ChainSafe/ChainBridge/message"
-	utils "github.com/ChainSafe/ChainBridge/utils/substrate"
-	subtest "github.com/ChainSafe/ChainBridge/utils/substrate/testing"
+	utils "github.com/ChainSafe/ChainBridge/shared/substrate"
+	subtest "github.com/ChainSafe/ChainBridge/shared/substrate/testing"
 	"github.com/centrifuge/go-substrate-rpc-client/types"
 )
