--- utils/ethereum/testing.go
@@ -1,4 +1,0 @@
-// Copyright 2020 ChainSafe Systems
-// SPDX-License-Identifier: LGPL-3.0-only
-
-package utils
