--- utils/logs.go
@@ -1,14 +1,0 @@
-// Copyright 2020 ChainSafe Systems
-// SPDX-License-Identifier: LGPL-3.0-only
-
-package utils
-
-import (
-	log "github.com/ChainSafe/log15"
-)
-
-func SetLogger(lvl log.Lvl) {
-	logger := log.Root()
-	handler := logger.GetHandler()
-	log.Root().SetHandler(log.LvlFilterHandler(lvl, handler))
-}
